customer separated to model and contr
customers now has a controller part for routing and a model part for interaction with the database. All seems to be working as before. this.changes can't be accessed outside callback, so manual get-calls to database has had to be made, to account for such checks instead. Otherwise the logic of functions should be the same as before.